fix: do not declare a script version for goatcounter (default to latest count.js)
defeats sha384 integrity check, but this was breaking all scripts on the
page for any user running a basic ad/tracking blocker, because astro
would hoist all scripts to the same running context such that any one
error in any of them would break all of them. you get a different sha384
for a blocked resource!